Create README.md
So this is my first mod that I am going to make and I really hope that you'll like it whoever will read this, and yeah have fun. oh also I do hope y'all don't see this until I'm done with the mod cuz it would be weird if 'yall could.